requirements.txt
aiohttp
appnope
beautifulsoup4
black
cachetools
certifi
chromadb
click
dataclasses-json
datasets
debugpy
decorator
dill
easyocr
faiss-cpu
fastapi
filelock
flask
google-auth
grpcio
huggingface-hub
ipykernel
ipython
joblib
jsonschema
langchain
matplotlib-inline
numpy
onnxruntime
openai
pandas
pillow
pre-commit
prompt-toolkit
pytest
requests
scikit-learn
scipy
sentence-transformers
streamlit
torch
torchaudio
torchvision
transformers
tqdm
uvicorn
watchfiles
pytesseract
instaloader
fake-useragent
langchain-groq
langchain-huggingface
langchain-community
python-dotenv
reportlab
aiohttp